Define navigation timings explicitly
Previously, we unmarshalled the return value of `performance.timings` as
a map of floats, which was rather brittle.

It seems some (more recent?) versions of Chrome include a `toJSON`
function as part of the `performance.timings` return value, which caused
unmarshaling to fail:

    ERRO[0036] failed to run script: unexpected response:
    {"sessionId":"acc66d616b5043208ed9d91c0e1084b4","status":0,"value":{"connectEnd":1468266125271,"connectStart":1468266125271,"domComplete":1468266126349,"domContentLoadedEventEnd":1468266125856,"domContentLoadedEventStart":1468266125701,"domInteractive":1468266125700,"domLoading":1468266125491,"domainLookupEnd":1468266125271,"domainLookupStart":1468266125271,"fetchStart":1468266125271,"loadEventEnd":1468266126350,"loadEventStart":1468266126349,"navigationStart":1468266125039,"redirectEnd":0,"redirectStart":0,"requestStart":1468266125293,"responseEnd":1468266125489,"responseStart":1468266125488,"secureConnectionStart":0,"toJSON":{},"unloadEventEnd":0,"unloadEventStart":0}}
    source=probe.go:56

Instead, be explicit about the exact values we want to parse and using a
struct and JSON tags.

This also removes the `snakeCase()` function which I wasn't keen on.

Fixes #1.